Changed Transaction amount to a decimal rather than a float so we can…

+++ src/Quibill.Domain/Models/ITransaction.cs
@@ -3,7 +3,7 @@
     internal interface ITransaction
     {
         int TransactionId { get; }
-        float TransactionAmount { get;}
+        decimal TransactionAmount { get;}
         string TransactionType { get; } // TODO make this an enum?
         System.DateTime AddDate { get; }
